src/webhook_service/trigger_webhook/requirements.txt
boto3
requests
pycryptodome